requirements.txt
Django==2.1.4
djangorestframework==3.9.2
mysqlclient==1.3.14
